[MOV] l10n_es_edi_facturae{_adm_centers}: merge modules
Since the administrative centers feature for the Spanish Facturae format
was added in stable, we needed to introduce a new "patch" module.

In master, the patch module can be merged into the main one.

task-3599447
