+++ addons/l10n_es_edi_facturae/security/ir.model.access.csv
@@ -1,3 +1,5 @@
 id,name,model_id/id,group_id/id,perm_read,perm_write,perm_create,perm_unlink
 l10n_es_edi_facturae.access_l10n_es_edi_facturae_certificate_manager,access_l10n_es_edi_facturae_certificate,l10n_es_edi_facturae.model_l10n_es_edi_facturae_certificate,account.group_account_manager,1,0,0,0
 l10n_es_edi_facturae.access_l10n_es_edi_facturae_certificate_system,access_l10n_es_edi_facturae_certificate,l10n_es_edi_facturae.model_l10n_es_edi_facturae_certificate,base.group_system,1,1,1,1
+l10n_es_edi_facturae.access_l10n_es_edi_facturae_ac_role_type_invoice,access_l10n_es_edi_facturae_ac_role_type,l10n_es_edi_facturae.model_l10n_es_edi_facturae_ac_role_type,account.group_account_invoice,1,0,0,0
+l10n_es_edi_facturae.access_l10n_es_edi_facturae_ac_role_type_readonly,access_l10n_es_edi_facturae_ac_role_type,l10n_es_edi_facturae.model_l10n_es_edi_facturae_ac_role_type,account.group_account_readonly,1,0,0,0
