--- addons/l10n_es_edi_facturae_adm_centers/__init__.py
@@ -1,1 +1,0 @@
-from . import models
