--- addons/l10n_es_edi_facturae_adm_centers/models/__init__.py
@@ -1,2 +1,0 @@
-from . import account_move
-from . import res_partner
