--- addons/l10n_es_edi_facturae_adm_centers/security/ir.model.access.csv
@@ -1,3 +1,0 @@
-id,name,model_id/id,group_id/id,perm_read,perm_write,perm_create,perm_unlink
-l10n_es_edi_facturae_adm_centers.access_l10n_es_edi_facturae_ac_role_type_invoice,access_l10n_es_edi_facturae_ac_role_type,l10n_es_edi_facturae_adm_centers.model_l10n_es_edi_facturae_adm_centers_ac_role_type,account.group_account_invoice,1,0,0,0
-l10n_es_edi_facturae_adm_centers.access_l10n_es_edi_facturae_ac_role_type_readonly,access_l10n_es_edi_facturae_ac_role_type,l10n_es_edi_facturae_adm_centers.model_l10n_es_edi_facturae_adm_centers_ac_role_type,account.group_account_readonly,1,0,0,0
