--- addons/l10n_es_edi_facturae_adm_centers/tests/__init__.py
@@ -1,1 +1,0 @@
-from . import test_edi_xml
